+++ docs/changes.rst
@@ -7,6 +7,7 @@ OpenREM version history
 --------------------
 * :issue:`960`  SkinDose: fixed bug where multiple entries in the openskin safelist table with different software versions caused an error
 * :issue:`958`  Interface: fixed task table sorting for Started column
+* :issue:`956`  Interface: enable updating of standard name database links
 * :issue:`955`  Imports: fixed error when importing RDSR with empty DAP measured value sequence
 * :issue:`953`  Interface: updated "DX and CR" to "Radiography", and "Radiographic" to "Radiography"
 * :issue:`950`  Documentation: removed (outdated) instructions for Conquest